refactor(provider-arch): phase 2 sprint 4 - codex app-server lifecycle extraction
Extract WebSocket lifecycle and process management functions from
runtimeCore.ts into two new codex/ modules.

codexAppServerClient.ts (~180 LOC):
- CodexAppServerSocket type
- buildCodexAppServerListenUrl
- reserveCodexAppServerPort
- createCodexAppServerSocket (internal)
- connectCodexAppServerSocket (retry loop, 5s timeout)
- normalizeCodexAppServerMessageData

codexAppServerLifecycle.ts (~180 LOC):
- buildCodexAppServerSpawnOptions
- launchDetachedCodexAppServerProcess (PID-reporting launcher bridge)
- terminateCodexAppServerProcess (process-group kill with fallback)
- rejectCodexAppServerPendingRequests
- createCodexAppServerAbortPromise

runtimeCore.ts: -9483 chars (10 functions + 1 type removed, 13-line
import block added).

Validation:
- tsc --noEmit clean (aris-backend).
- vitest run --exclude e2e: 301/301 PASS (regression 0).

Refs Phase 2 Sprint 4 of
docs/03-platform/codex-backend-alignment-plan.md.
