+++ services/aris-backend/src/runtime/runtimeCore.ts
@@ -31,6 +31,20 @@ import { GeminiStreamAdapter } from './providers/gemini/geminiStreamAdapter.js';
 import { GeminiSessionRegistry } from './providers/gemini/geminiSessionRegistry.js';
 import { buildProviderCommand, type ProviderCommand } from './providers/providerCommandFactory.js';
 import {
+  buildCodexAppServerListenUrl,
+  connectCodexAppServerSocket,
+  normalizeCodexAppServerMessageData,
+  reserveCodexAppServerPort,
+  type CodexAppServerSocket,
+} from './providers/codex/codexAppServerClient.js';
+import {
+  buildCodexAppServerSpawnOptions,
+  createCodexAppServerAbortPromise,
+  launchDetachedCodexAppServerProcess,
+  rejectCodexAppServerPendingRequests,
+  terminateCodexAppServerProcess,
+} from './providers/codex/codexAppServerLifecycle.js';
+import {
   buildCodexPermissionKey,
   buildCodexThreadCacheKey,
   classifyCodexAppServerFailure,
@@ -1179,309 +1193,6 @@ function buildAgentCommand(
   });
 }
 
-function buildCodexAppServerSpawnOptions(input: {
-  cwd: string;
-  env: NodeJS.ProcessEnv;
-  signal?: AbortSignal;
-}): SpawnOptionsWithoutStdio {
-  return {
-    cwd: input.cwd,
-    env: input.env,
-    stdio: 'pipe',
-    signal: input.signal,
-    detached: true,
-  };
-}
-
-function buildCodexAppServerListenUrl(port: number): string {
-  return `ws://127.0.0.1:${port}`;
-}
-
-async function reserveCodexAppServerPort(): Promise<number> {
-  return new Promise<number>((resolve, reject) => {
-    const server = createServer();
-    server.once('error', reject);
-    server.listen(0, '127.0.0.1', () => {
-      const address = server.address();
-      if (!address || typeof address === 'string') {
-        server.close();
-        reject(new Error('failed to reserve codex app-server websocket port'));
-        return;
-      }
-      const port = address.port;
-      server.close((error) => {
-        if (error) {
-          reject(error);
-          return;
-        }
-        resolve(port);
-      });
-    });
-    server.unref?.();
-  });
-}
-
-type CodexAppServerSocket = {
-  readonly readyState: number;
-  send(data: string): void;
-  close(code?: number, reason?: string): void;
-  addEventListener(type: 'open', listener: () => void): void;
-  addEventListener(type: 'message', listener: (event: { data: unknown }) => void): void;
-  addEventListener(type: 'error', listener: (event: unknown) => void): void;
-  addEventListener(type: 'close', listener: () => void): void;
-  removeEventListener(type: 'open', listener: () => void): void;
-  removeEventListener(type: 'message', listener: (event: { data: unknown }) => void): void;
-  removeEventListener(type: 'error', listener: (event: unknown) => void): void;
-  removeEventListener(type: 'close', listener: () => void): void;
-};
-
-function createCodexAppServerSocket(url: string): CodexAppServerSocket {
-  const WebSocketCtor = (globalThis as { WebSocket?: new (url: string) => CodexAppServerSocket }).WebSocket;
-  if (typeof WebSocketCtor !== 'function') {
-    throw new Error('global WebSocket constructor is unavailable');
-  }
-  return new WebSocketCtor(url);
-}
-
-async function connectCodexAppServerSocket(
-  url: string,
-  options: { signal?: AbortSignal; timeoutMs?: number } = {},
-): Promise<CodexAppServerSocket> {
-  const timeoutMs = options.timeoutMs ?? 5_000;
-  const deadline = Date.now() + timeoutMs;
-  let lastError: Error | null = null;
-
-  while (Date.now() < deadline) {
-    if (options.signal?.aborted) {
-      throw new Error('codex app-server websocket connection aborted');
-    }
-
-    try {
-      const socket = await new Promise<CodexAppServerSocket>((resolve, reject) => {
-        const candidate = createCodexAppServerSocket(url);
-        let settled = false;
-
-        const cleanup = () => {
-          candidate.removeEventListener('open', handleOpen);
-          candidate.removeEventListener('error', handleError);
-          candidate.removeEventListener('close', handleClose);
-          options.signal?.removeEventListener('abort', handleAbort);
-          clearTimeout(timer);
-        };
-
-        const finishResolve = () => {
-          if (settled) {
-            return;
-          }
-          settled = true;
-          cleanup();
-          resolve(candidate);
-        };
-
-        const finishReject = (error: Error) => {
-          if (settled) {
-            return;
-          }
-          settled = true;
-          cleanup();
-          try {
-            candidate.close();
-          } catch {
-            // ignore close failures while cleaning up failed attempts
-          }
-          reject(error);
-        };
-
-        const handleOpen = () => finishResolve();
-        const handleError = () => finishReject(new Error('codex app-server websocket connection failed'));
-        const handleClose = () => finishReject(new Error('codex app-server websocket closed before opening'));
-        const handleAbort = () => finishReject(new Error('codex app-server websocket connection aborted'));
-        const timer = setTimeout(() => finishReject(new Error('timed out waiting for codex app-server websocket')), 750);
-
-        candidate.addEventListener('open', handleOpen);
-        candidate.addEventListener('error', handleError);
-        candidate.addEventListener('close', handleClose);
-        options.signal?.addEventListener('abort', handleAbort, { once: true });
-      });
-      return socket;
-    } catch (error) {
-      lastError = error instanceof Error ? error : new Error(String(error));
-      if (Date.now() >= deadline) {
-        break;
-      }
-      await new Promise((resolve) => setTimeout(resolve, 50));
-    }
-  }
-
-  throw lastError ?? new Error('failed to connect to codex app-server websocket');
-}
-
-function normalizeCodexAppServerMessageData(data: unknown): string {
-  if (typeof data === 'string') {
-    return data;
-  }
-  if (data instanceof ArrayBuffer) {
-    return Buffer.from(data).toString('utf8');
-  }
-  if (ArrayBuffer.isView(data)) {
-    return Buffer.from(data.buffer, data.byteOffset, data.byteLength).toString('utf8');
-  }
-  return String(data ?? '');
-}
-
-async function launchDetachedCodexAppServerProcess(input: {
-  cwd: string;
-  env: NodeJS.ProcessEnv;
-  args: string[];
-  signal?: AbortSignal;
-}): Promise<number> {
-  const launcherScript = `
-    const { spawn } = require('node:child_process');
-    const command = process.argv[1];
-    const args = JSON.parse(process.argv[2]);
-    const cwd = process.argv[3];
-    try {
-      const child = spawn(command, args, {
-        cwd,
-        env: process.env,
-        detached: true,
-        stdio: 'ignore',
-      });
-      if (!child.pid || !Number.isInteger(child.pid) || child.pid <= 0) {
-        console.error('missing detached codex app-server pid');
-        process.exit(1);
-      }
-      console.log(String(child.pid));
-      child.unref();
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      console.error(message);
-      process.exit(1);
-    }
-  `;
-  const launcher = spawn(process.execPath, ['-e', launcherScript, 'codex', JSON.stringify(input.args), input.cwd], {
-    cwd: input.cwd,
-    env: input.env,
-    stdio: ['ignore', 'pipe', 'pipe'],
-    signal: input.signal,
-  });
-  const stdoutChunks: string[] = [];
-  const stderrChunks: string[] = [];
-  launcher.stdout?.on('data', (chunk: Buffer | string) => {
-    stdoutChunks.push(typeof chunk === 'string' ? chunk : chunk.toString('utf8'));
-  });
-  launcher.stderr?.on('data', (chunk: Buffer | string) => {
-    stderrChunks.push(typeof chunk === 'string' ? chunk : chunk.toString('utf8'));
-  });
-
-  const exitCode = await new Promise<number | null>((resolve, reject) => {
-    launcher.once('error', reject);
-    launcher.once('close', (code) => resolve(code));
-  });
-  if (exitCode !== 0) {
-    const detail = stderrChunks.join('').trim() || stdoutChunks.join('').trim() || `exit code ${exitCode ?? 'null'}`;
-    throw new Error(`failed to launch detached codex app-server: ${detail}`);
-  }
-
-  const pid = Number.parseInt(stdoutChunks.join('').trim(), 10);
-  if (!Number.isInteger(pid) || pid <= 0) {
-    throw new Error(`failed to parse detached codex app-server pid: ${stdoutChunks.join('').trim()}`);
-  }
-  return pid;
-}
-
-function terminateCodexAppServerProcess(
-  child: { pid?: number; killed?: boolean; kill: (signal?: NodeJS.Signals | number) => boolean },
-  killProcess: (pid: number, signal?: NodeJS.Signals | number) => void = process.kill,
-  signal: NodeJS.Signals = 'SIGTERM',
-): void {
-  if (child.killed) {
-    return;
-  }
-
-  const pid = typeof child.pid === 'number' && Number.isInteger(child.pid) && child.pid > 0
-    ? child.pid
-    : null;
-  if (pid !== null) {
-    try {
-      killProcess(-pid, signal);
-      return;
-    } catch {
-      // Fall back to the direct child signal if the process group no longer exists.
-    }
-  }
-
-  child.kill(signal);
-}
-
-function rejectCodexAppServerPendingRequests(
-  pendingRequests: Map<string, {
-    method: string;
-    reject: (error: Error) => void;
-  }>,
-  reason: string,
-): void {
-  for (const [key, pending] of pendingRequests.entries()) {
-    pending.reject(new Error(reason));
-    pendingRequests.delete(key);
-  }
-}
-
-function createCodexAppServerAbortPromise(input: {
-  signal?: AbortSignal;
-  pendingRequests: Map<string, {
-    method: string;
-    reject: (error: Error) => void;
-  }>;
-  onAbort: () => void;
-}): {
-  interrupted: Promise<{ status: 'interrupted' }> | null;
-  dispose: () => void;
-} {
-  if (!input.signal) {
-    return {
-      interrupted: null,
-      dispose: () => {},
-    };
-  }
-
-  let disposed = false;
-  let handleAbort: (() => void) | null = null;
-  const finalizeAbort = (resolve: (value: { status: 'interrupted' }) => void) => {
-    if (disposed) {
-      return;
-    }
-    disposed = true;
-    rejectCodexAppServerPendingRequests(input.pendingRequests, 'The operation was aborted');
-    input.onAbort();
-    resolve({ status: 'interrupted' });
-  };
-
-  const interrupted = new Promise<{ status: 'interrupted' }>((resolve) => {
-    if (input.signal?.aborted) {
-      finalizeAbort(resolve);
-      return;
-    }
-
-    const abortListener = () => {
-      input.signal?.removeEventListener('abort', abortListener);
-      finalizeAbort(resolve);
-    };
-    handleAbort = abortListener;
-
-    input.signal?.addEventListener('abort', abortListener, { once: true });
-  });
-
-  return {
-    interrupted,
-    dispose: () => {
-      disposed = true;
-      if (handleAbort) {
-        input.signal?.removeEventListener('abort', handleAbort);
-      }
-    },
-  };
-}
 
 export const runtimeCoreTestHooks = {
   parseAgentStreamLine,
